+++ apps/commons/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CommonsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'commons'
+    name = 'apps.commons'
