+++ config/settings.py
@@ -1,15 +1,3 @@
-"""
-Django settings for config project.
-
-Generated by 'django-admin startproject' using Django 5.0.4.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/5.0/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/5.0/ref/settings/
-"""
-
 from pathlib import Path
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
@@ -30,9 +18,9 @@ ALLOWED_HOSTS = ['*']
 AUTH_USER_MODEL = 'users.User'
 
 LOCAL_APPS = [
-    'users',
-    'api',
-    'commons',
+    'apps.users.apps.UsersConfig',
+    'apps.api.apps.ApiConfig',
+    'apps.commons.apps.CommonsConfig',
 ]
 
 THIRD_PARTH_APPS = [
@@ -135,8 +123,8 @@ DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 REST_FRAMEWORK = {
     'DEFAULT_RENDERER_CLASSES': [
-        # 'rest_framework.renderers.BrowsableAPIRenderer',
-        'commons.renderers.CustomJSONRenderer',
+        'rest_framework.renderers.BrowsableAPIRenderer',
+        'apps.commons.renderers.CustomJSONRenderer',
         # 'rest_framework.renderers.JSONRenderer',
     ],
     'DEFAULT_PARSER_CLASSES': [
