+++ config/urls.py
@@ -17,8 +17,8 @@ Including another URLconf
 # from django.contrib import admin
 from django.urls import path, include
 
-from users import urls as users
-from api import urls as apis
+from apps.users import urls as users
+from apps.api import urls as apis
 
 
 urlpatterns = [
